requirements.txt
# NHANES Periodontitis ML Project
# Requirements

# Data handling
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # For parquet files

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
catboost>=1.2.0
lightgbm>=4.0.0
optuna>=3.3.0  # Hyperparameter optimization

# Interpretability
shap>=0.43.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Statistics
scipy>=1.11.0
statsmodels>=0.14.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0.0  # For config.yaml loading

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
